+++ CHANGELOG.md
@@ -14,6 +14,8 @@ The changelog is applicable from version `1.0.0` onwards.
 
 ### Added
 
+- APPCLI-130: Use Devbox to manage developer environment.
+
 ### Changed
 
 ### Deprecated
@@ -75,7 +77,6 @@ The changelog is applicable from version `1.0.0` onwards.
 ### Security
 
 - APPCLI-128: Patch [jinja2 attr vulnerability](https://github.com/pallets/jinja/security/advisories/GHSA-cpwx-vrp4-4pq7)
-
 
 ---