+++ CHANGELOG.md
@@ -18,6 +18,8 @@ The changelog is applicable from version `1.0.0` onwards.
 
 ### Changed
 
+- TERA-859: Expose orchestrator from the configuration to the CLI context.
+
 ### Deprecated
 
 ### Removed