+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The changelog is applicable from version `1.0.0` onwards.
 ### Added
 
 - APPCLI-130: Use Devbox to manage developer environment.
+- APPCLI-137: Include code quality tooling.
 
 ### Changed