+++ CHANGELOG.md
@@ -19,6 +19,8 @@ The changelog is applicable from version `1.0.0` onwards.
 
 ### Changed
 
+- NSWCC-752: Update Appcli image name in readme and quickstart.md.
+
 ### Deprecated
 
 ### Removed